requirements.txt
flask
pymysql
pandas
requests
beautifulsoup4
feedparser
python-dotenv
cloudscraper